PicQuick
"Pic Quick" helps you to figure out what is the item and who made it. The user could use any of the android device and install the app. You just simply snap the item, and Pic Quick will scan the picture for hidden logos and take it features and search for items on the internet that have the qualities. It also gives you access to the item on online stores (such as Amazon or Flipkart). If Pic Quick is not sure about the creator, it will just return the item name.